ux z bardzo slabego na slaby

+++ Wypozyczalnia/Formsy/FrmEdyfujTaryfe.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -16,17 +17,8 @@ namespace Wypozyczalnia.Formsy
         string connString = "Data Source = baza.db; Version = 3";
         int id;
 
-        private void btn_zatwierdz_control()
-        {
-            if (string.IsNullOrWhiteSpace(txt_cena_za_dzien.Text) || string.IsNullOrWhiteSpace(txt_nazwa.Text)) btn_zatwierdz.Enabled = false;
-            else btn_zatwierdz.Enabled = true;
-        }
-
         private void btn_dodaj_control()
         {
-            if (String.IsNullOrWhiteSpace(txt_dni.Text) || String.IsNullOrWhiteSpace(txt_cena.Text)) btn_dodaj.Enabled = false;
-            else btn_dodaj.Enabled = true;
-
             foreach (string s in lb_zasady.Items)
             {
                 if (s.StartsWith(txt_dni.Text))
@@ -82,19 +74,22 @@ namespace Wypozyczalnia.Formsy
 
         private void btn_dodaj_Click(object sender, EventArgs e)
         {
-            for (int i = 0; i < lb_zasady.Items.Count; i++)
+            if (!String.IsNullOrWhiteSpace(txt_dni.Text) && !String.IsNullOrWhiteSpace(txt_cena.Text))
             {
-                if (lb_zasady.Items[i].ToString().StartsWith(txt_dni.Text))
+                for (int i = 0; i < lb_zasady.Items.Count; i++)
                 {
-                    lb_zasady.Items[i] = txt_dni.Text + " dni - " + txt_cena.Text + "zł";
-                    lb_zasady.SelectedItem = txt_dni.Text + " dni - " + txt_cena.Text + "zł";
-                    btn_dodaj_control();
-                    return;
+                    if (lb_zasady.Items[i].ToString().StartsWith(txt_dni.Text))
+                    {
+                        lb_zasady.Items[i] = txt_dni.Text + " dni - " + txt_cena.Text + "zł";
+                        lb_zasady.SelectedItem = txt_dni.Text + " dni - " + txt_cena.Text + "zł";
+                        btn_dodaj_control();
+                        return;
+                    }
                 }
+                lb_zasady.Items.Add(txt_dni.Text + " dni - " + txt_cena.Text + "zł");
+                lb_zasady.SelectedItem = txt_dni.Text + " dni - " + txt_cena.Text + "zł";
+                btn_dodaj_control();
             }
-            lb_zasady.Items.Add(txt_dni.Text + " dni - " + txt_cena.Text + "zł");
-            lb_zasady.SelectedItem = txt_dni.Text + " dni - " + txt_cena.Text + "zł";
-            btn_dodaj_control();
         }
 
         private void btn_usun_Click(object sender, EventArgs e)
@@ -147,9 +142,7 @@ namespace Wypozyczalnia.Formsy
                 string[] words = lb_zasady.SelectedItem.ToString().Split(' ');
                 txt_dni.Text = words[0];
                 txt_cena.Text = words[3].Trim(new Char[] { 'z', 'ł' });
-                btn_usun.Enabled = true;
             }
-            else btn_usun.Enabled = false;
         }
 
         private void txt_cena_za_dzien_KeyPress(object sender, KeyPressEventArgs e)
@@ -160,15 +153,8 @@ namespace Wypozyczalnia.Formsy
             }
         }
 
-        private void txt_nazwa_TextChanged(object sender, EventArgs e)
-        {
-            btn_zatwierdz_control();
-        }
-
         private void txt_cena_za_dzien_TextChanged(object sender, EventArgs e)
         {
-            btn_zatwierdz_control();
-
             string temp = null;
             foreach (char c in txt_cena_za_dzien.Text)
             {
@@ -187,45 +173,49 @@ namespace Wypozyczalnia.Formsy
 
         private void btn_zatwierdz_Click(object sender, EventArgs e)
         {
-            using (SQLiteConnection conn = new SQLiteConnection(connString))
+            if (!string.IsNullOrWhiteSpace(txt_cena_za_dzien.Text) && !string.IsNullOrWhiteSpace(txt_nazwa.Text))
             {
-                conn.Open();
+                using (SQLiteConnection conn = new SQLiteConnection(connString))
+                {
+                    conn.Open();
 
-                SQLiteCommand command1 = new SQLiteCommand(conn);
-                command1.CommandText = @"
+                    SQLiteCommand command1 = new SQLiteCommand(conn);
+                    command1.CommandText = @"
                                 UPDATE Taryfy
                                 SET
                                 nazwa = @nazwa,
                                 cena_za_dzien = @cena_za_dzien
                                 WHERE id = @id
                                 ";
-                command1.Parameters.Add(new SQLiteParameter("@id", id));
-                command1.Parameters.Add(new SQLiteParameter("@nazwa", txt_nazwa.Text));
-                command1.Parameters.Add(new SQLiteParameter("@cena_za_dzien", txt_cena_za_dzien.Text));                
-                command1.ExecuteNonQuery();
+                    command1.Parameters.Add(new SQLiteParameter("@id", id));
+                    command1.Parameters.Add(new SQLiteParameter("@nazwa", txt_nazwa.Text));
+                    command1.Parameters.Add(new SQLiteParameter("@cena_za_dzien", txt_cena_za_dzien.Text));
+                    command1.ExecuteNonQuery();
 
-                SQLiteCommand command2 = new SQLiteCommand(conn);
-                command2.CommandText = "DELETE FROM TaryfyZasady WHERE id = @id";
-                command2.Parameters.Add(new SQLiteParameter("@id", id));
-                command2.ExecuteNonQuery();
+                    SQLiteCommand command2 = new SQLiteCommand(conn);
+                    command2.CommandText = "DELETE FROM TaryfyZasady WHERE id = @id";
+                    command2.Parameters.Add(new SQLiteParameter("@id", id));
+                    command2.ExecuteNonQuery();
 
-                foreach (string s in lb_zasady.Items)
-                {
-                    string[] words = s.Split(' ');
-                    SQLiteCommand command3 = new SQLiteCommand(conn);
-                    command3.CommandText = @"
+                    foreach (string s in lb_zasady.Items)
+                    {
+                        string[] words = s.Split(' ');
+                        SQLiteCommand command3 = new SQLiteCommand(conn);
+                        command3.CommandText = @"
                                 INSERT INTO TaryfyZasady (id, dni, cena)
                                 VALUES (@id, @dni, @cena);
                                 ;";
-                    command3.Parameters.Add(new SQLiteParameter("@id", id));
-                    command3.Parameters.Add(new SQLiteParameter("@dni", words[0]));
-                    command3.Parameters.Add(new SQLiteParameter("@cena", words[3]));
-                    command3.ExecuteNonQuery();
-                }
+                        command3.Parameters.Add(new SQLiteParameter("@id", id));
+                        command3.Parameters.Add(new SQLiteParameter("@dni", words[0]));
+                        command3.Parameters.Add(new SQLiteParameter("@cena", words[3]));
+                        command3.ExecuteNonQuery();
+                    }
 
-                conn.Close();
-                this.Close();
+                    conn.Close();
+                    this.Close();
+                }
             }
+            else MessageBox.Show("Uzupełnij wymagane pola", "Uwaga", MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
         }
     }
 }
